Add untested config management functionality
Definitely will want some property tests for the name generation, not
convinced as written that it will function as written.

Might need to shortcircuit the retry logic if the maximum pairs of names
that can be generated is reached. Would be a weird situation to be in
but ya never know.

Error handling still sucks, and I'm not totally sold on the
implementation so far. Maybe some testing will reveal a more elegant
solution. Also starting to get concerned with how hard it is to pilot new
functionality. Should consider adding some test data to a project-level
ghci config file to make that easier.