Add top-level edid module, add Generic instances to Display
Top-level package is to simplify edid usage. Display Generic instances
is to simplify smallcheck series generation.

I'm kind of bummed that the display depth has to be set so low for
Display objects now. I'm not sure what the fix is, but I'll think about
it sometime down the road.